# QuantDinger-main/docs/api/openapi.yaml
info:
  description: 'Human-facing REST API for QuantDinger. Agent integrations use `/api/agent/v1`
    — see `docs/agent/agent-openapi.json`. Conventions: `docs/architecture/API_CONVENTIONS.md`.'
  contact:
    name: QuantDinger
    url: https://github.com/OpenByteInc/quantdinger
  license:
    name: See repository LICENSE
  title: QuantDinger Web API
  version: 1.0.0
components:
  securitySchemes:
    HumanJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'User session JWT from `POST /api/auth/login`. Send as `Authorization:
        Bearer <token>`.'
    AgentToken:
      type: http
      scheme: bearer
      bearerFormat: qd_agent_xxx
      description: Agent Gateway token for `/api/agent/v1/*` only. See docs/agent/AGENT_QUICKSTART.md.
  schemas:
...
    HumanSuccessEnvelope:
      type: object
      properties:
        code:
          type: integer
          description: 1 on success
          example: 1
        msg:
          type: string
          example: success
        data:
          nullable: true
      additionalProperties: false
    HumanErrorEnvelope:
      type: object
      properties:
        code:
          type: integer
          description: 0 on error
          example: 0
        msg:
          type: string
        data:
          nullable: true
      additionalProperties: false
...
servers:
- url: http://localhost:5000
  description: Backend direct (python run.py)
- url: http://localhost:8888
  description: Docker Compose (via reverse proxy)
paths:
...
  /api/community/indicators:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: List published marketplace indicators.
      description: "Query params:\npage: Page number (default 1)\npage_size: Page\
        \ size (default 12)\nkeyword: Search keyword\npricing_type: 'free' / 'paid'\
        \ / empty (all)\nvip_free: 1/true to show VIP-free paid assets only\ncode_visibility:\
        \ 'visible' / 'hidden' / empty (all)\nmin_price, max_price: Numeric price\
        \ bounds\nsort_by: 'score' (default) / 'newest' / 'hot' / 'price_asc' /\n\
        \         'price_desc' / 'rating'.\n         'score' sorts by the composite\
        \ Strategy API V2 backtest score\n         (see services/community_kpis.py)\
        \ and is now the\n         default — putting genuinely well-performing indicators\n\
        \         at the top of the marketplace, not just the newest ones."
      tags:
      - Community
      operationId: getApiCommunityIndicators
      x-visibility: public
  /api/community/indicators/{indicator_id}:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get marketplace indicator detail.
      tags:
      - Community
      operationId: getApiCommunityIndicatorsIndicatorId
      parameters:
      - name: indicator_id
        in: path
        required: true
        schema:
          type: integer
      x-visibility: public
    parameters:
    - in: path
      name: indicator_id
      required: true
      schema:
        type: integer
        minimum: 0
  /api/community/indicators/{indicator_id}/compatibility:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      tags:
      - Community
      operationId: getApiCommunityIndicatorsIndicatorIdCompatibility
      summary: Get Api Community Indicators Indicator Id Compatibility
      parameters:
      - name: indicator_id
        in: path
        required: true
        schema:
          type: integer
      x-visibility: public
    parameters:
    - in: path
      name: indicator_id
      required: true
      schema:
        type: integer
        minimum: 0
  /api/community/indicators/{indicator_id}/adapt:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      tags:
      - Community
      operationId: postApiCommunityIndicatorsIndicatorIdAdapt
      summary: Post Api Community Indicators Indicator Id Adapt
      parameters:
      - name: indicator_id
        in: path
        required: true
        schema:
          type: integer
      x-visibility: public
    parameters:
    - in: path
      name: indicator_id
      required: true
      schema:
        type: integer
        minimum: 0
  /api/community/indicators/{indicator_id}/purchase:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Purchase a marketplace indicator.
      description: 'Side effects:

        1. Verify buyer has enough credits

        2. Debit buyer credits and credit seller

        3. Create purchase record

        4. Copy indicator into buyer account'
      tags:
      - Community
      operationId: postApiCommunityIndicatorsIndicatorIdPurchase
      parameters:
      - name: indicator_id
        in: path
        required: true
        schema:
          type: integer
      x-visibility: public
    parameters:
    - in: path
      name: indicator_id
      required: true
      schema:
        type: integer
        minimum: 0
  /api/community/indicators/{indicator_id}/sync:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Sync latest code for a purchased indicator.
      description: "Use when the publisher updated code after the buyer purchased;\
        \ the buyer\npulls the newest version into their local copy.\n\nPreconditions:\n\
        \    - Caller must have purchased the indicator\n    - Published listings\
        \ sync from the latest marketplace code\n    - Unpublished/deleted listings\
        \ can still restore from the purchase snapshot"
      tags:
      - Community
      operationId: postApiCommunityIndicatorsIndicatorIdSync
      parameters:
      - name: indicator_id
        in: path
        required: true
        schema:
          type: integer
      x-visibility: public
    parameters:
    - in: path
      name: indicator_id
      required: true
      schema:
        type: integer
        minimum: 0
  /api/community/my-purchases:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: List indicators purchased by the current user.
      tags:
      - Community
      operationId: getApiCommunityMyPurchases
      x-visibility: public
  /api/community/author/summary:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: 'Author overview: published / approved / pending / sales / revenue
        / avg rating.'
      tags:
      - Community
      operationId: getApiCommunityAuthorSummary
      x-visibility: public
  /api/community/author/published:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get Api Community Author Published
      tags:
      - Community
      operationId: getApiCommunityAuthorPublished
      description: List indicators published by the author with per-item sales, revenue,
        and rating.
      x-visibility: public
  /api/community/author/sales:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Paginated author sales ledger; optional indicator_id filter.
      tags:
      - Community
      operationId: getApiCommunityAuthorSales
      x-visibility: public
  /api/community/author/indicators/{indicator_id}/unpublish:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Let an author unpublish their own marketplace asset.
      tags:
      - Community
      operationId: postApiCommunityAuthorIndicatorsIndicatorIdUnpublish
      parameters:
      - name: indicator_id
        in: path
        required: true
        schema:
          type: integer
      x-visibility: public
    parameters:
    - in: path
      name: indicator_id
      required: true
      schema:
        type: integer
        minimum: 0
...
  /api/community/indicators/{indicator_id}/my-comment:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get current user's comment on an indicator (for edit form).
      tags:
      - Community
      operationId: getApiCommunityIndicatorsIndicatorIdMyComment
      parameters:
      - name: indicator_id
        in: path
        required: true
        schema:
          type: integer
      x-visibility: public
    parameters:
    - in: path
      name: indicator_id
      required: true
      schema:
        type: integer
        minimum: 0
...
openapi: 3.0.3